BUGFIX : broadcasting error in optimize
AgeReplacementPolicy (+ OneCycle...) may have float-like costs but
encapsulate many assets. Then x0 is a float but eq(x0) used in
scipy.newton is an array. But scipy.newton expects a function taking a
shape and returning the same shape.

Solution : infer the shape of sf to get the number of underlying assets.