PR revision
- time_window : I propose a small refactoring of the interface. With t0
optional and default to 0, its position in the arguments is not
intuitive for the user. Passing a tuple of (t0, tf) reduce the number of
arguments and make request more explicit
- important : use black and isort formatter
- remove useless keyword arguments : I prefer not to use keyword
arguments if the number of arguments is limited because it encourages to
write function call with different order of arguments. Then the code is
less consistent. It is only a personal opinion. In this case, I think
the number of arguments is small enough not to use keyword arguments in
place of positional arguments.
- refactoring get_lifetime_model_nb_assets : this belongs to utils. To
avoid import errors, import in the function scope. I've also renamed the
function to get_model_nb_assets as it is not specific to lifetime model
and I've make it more generic so that it can be applied on any model
 - clean import : don't write from
relife.lifetime_model.conditional_model import ... Prefer from
relife.lifetime_model import ... as the exposition of the public
interface is held by lifetime_model. I think it might be better to hide
lifetime_model modules to avoid this kind of imports
- remove unused type hints
- missing __all__ to specify the exposed API
- hide methods that are not supposed to be part of the public object
interface but only belong to internal functionalities
- hide iterator : not exposed publically, only iterables are used
- remove step method : ambiguous with sample_step. I've also renamed
sample_step to sample_time_event_entry as its purpose is to sample time,
event, entry. It is more explicit.
- refactoring of get_rvs_size : not specific to every
StochasticDataIterator but only to those using a lifetime model. I've
removed the method from the parent interface and propose to use specific
properties for each derived class (if needed)